# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Aquarium Controller API
  description: |
    REST API untuk Smart Aquarium Controller dengan fitur Automatic Feeder dan UV Sterilizer.

    ## Fitur Utama
    - **Automatic Feeder**: Sistem pemberian pakan dengan mekanisme Double Gate
    - **UV Sterilizer**: Kontrol lampu UV dengan penjadwalan dan manual override
    - **Scheduler**: Penjadwalan otomatis untuk feeding dan UV
    - **History**: Log semua aktivitas device

    ## Demo Mode
    API ini mendukung demo mode yang mensimulasikan device responses tanpa memerlukan MQTT broker atau hardware.
  version: 1.0.0
  contact:
    name: API Support
    email: support@aquarium.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://iot-backend-cursor-production.up.railway.app/api/v1
    description: Local development server (default)
  - url: https://iot-backend-cursor-production.up.railway.app/api/v1
    description: Production server
...
paths:
...
  /feeder/schedules:
    get:
      tags:
        - Feeder
      summary: Get all feeder schedules
      description: Mengambil semua jadwal pakan yang telah dikonfigurasi dengan pagination
      operationId: getFeederSchedules
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
      responses:
        "200":
          description: List jadwal pakan dengan pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PakanSchedule"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      tags:
        - Feeder
      summary: Create new feeder schedule
      description: Membuat jadwal pakan baru. Maksimal 5 jadwal per hari.
      operationId: createFeederSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PakanScheduleInput"
            example:
              day_name: "Monday"
              time: "08:00"
              amount_gram: 10
              is_active: true
      responses:
        "201":
          description: Jadwal berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PakanSchedule"
        "400":
          description: Bad request (misal sudah ada 5 jadwal di hari tersebut)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /feeder/schedules/{id}:
    put:
      tags:
        - Feeder
      summary: Update feeder schedule
      description: Update jadwal pakan yang sudah ada
      operationId: updateFeederSchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID jadwal pakan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PakanScheduleInput"
      responses:
        "200":
          description: Jadwal berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PakanSchedule"
        "404":
          description: Jadwal tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Feeder
      summary: Delete feeder schedule
      description: Menghapus jadwal pakan
      operationId: deleteFeederSchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID jadwal pakan
      responses:
        "200":
          description: Jadwal berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Schedule deleted"
        "404":
          description: Jadwal tidak ditemukan
  /feeder/manual:
    post:
      tags:
        - Feeder
      summary: Trigger manual feed
      description: |
        Memicu pemberian pakan secara manual. 
        Sistem akan langsung mengurangi stock tanpa delay.
        User dapat menentukan jumlah pakan dalam gram (default 10 gram jika tidak diisi).
      operationId: manualFeed
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                amount_gram:
                  type: integer
                  minimum: 1
                  description: Jumlah pakan dalam gram (kelipatan 10 gram per dosis)
                  example: 15
            example:
              amount_gram: 15
      responses:
        "200":
          description: Command berhasil dikirim
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feeding command sent"
                  action_id:
                    type: integer
                    example: 26
                  amount_gram:
                    type: integer
                    example: 15
                  last_feed:
                    $ref: "#/components/schemas/LastFeedInfo"
              example:
                message: "Feeding command sent"
                action_id: 26
                amount_gram: 15
                last_feed:
                  day: "Wednesday"
                  time: "08:00"
...
  /history:
    get:
      tags:
        - History
      summary: Get action history
      description: Mengambil history aktivitas dengan optional filtering dan pagination
      operationId: getHistory
      parameters:
        - name: device_type
          in: query
          schema:
            type: string
            enum: [FEEDER, UV]
          description: Filter by device type
        - name: trigger_source
          in: query
          schema:
            type: string
            enum: [SCHEDULE, MANUAL]
          description: Filter by trigger source
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, RUNNING, SUCCESS, FAILED, STOPPED]
          description: Filter by status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of items per page (max 200)
      responses:
        "200":
          description: List history dengan pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ActionHistory"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
  /stock:
    get:
      tags:
        - Stock
      summary: Get current stock
      description: Mengambil jumlah stock pakan saat ini
      operationId: getStock
      responses:
        "200":
          description: Stock data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
    put:
      tags:
        - Stock
      summary: Update stock
      description: Update jumlah stock pakan
      operationId: updateStock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount_gram
              properties:
                amount_gram:
                  type: integer
                  minimum: 0
                  description: Jumlah stock dalam gram
                  example: 2000
      responses:
        "200":
          description: Stock berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
        "400":
          description: Bad request (amount negatif)
  /sensors/current:
    get:
      tags:
        - Sensors
      summary: Get current sensor readings
      description: Mengambil pembacaan sensor terbaru (temperature dan humidity)
      operationId: getCurrentSensor
      responses:
        "200":
          description: Current sensor data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorLog"
              example:
                id: 1
                temperature: 27.5
                humidity: 65.0
                recorded_at: "2025-11-19T09:30:00Z"
        "404":
          description: No sensor data available
  /sensors/history:
    get:
      tags:
        - Sensors
      summary: Get sensor history
      description: Mengambil history pembacaan sensor dengan pagination dan filter time range
      operationId: getSensorHistory
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of items per page (max 200)
        - name: range
          in: query
          schema:
            type: string
            enum: [24h, 7d, 30d]
          description: Time range filter (24h = last 24 hours, 7d = last 7 days, 30d = last 30 days)
      responses:
        "200":
          description: Paginated sensor history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SensorLog"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
              example:
                data:
                  - id: 100
                    temperature: 27.5
                    humidity: 65.0
                    recorded_at: "2025-11-19T09:30:00Z"
                  - id: 99
                    temperature: 27.3
                    humidity: 64.5
                    recorded_at: "2025-11-19T09:25:00Z"
                pagination:
                  page: 1
                  page_size: 50
                  total: 250
                  total_pages: 5
...
components:
  schemas:
    PakanSchedule:
      type: object
      properties:
        id:
          type: integer
          example: 1
        day_name:
          type: string
          enum: [Mon, Tue, Wed, Thu, Fri, Sat, Sun]
          example: "Mon"
        time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          description: Waktu dalam format HH:MM
          example: "08:00"
        amount_gram:
          type: integer
          minimum: 1
          default: 10
          description: Jumlah pakan dalam gram
          example: 10
        is_active:
          type: boolean
          default: true
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PakanScheduleInput:
      type: object
      required:
        - day_name
        - time
      properties:
        day_name:
          type: string
          enum: [Mon, Tue, Wed, Thu, Fri, Sat, Sun]
        time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
        amount_gram:
          type: integer
          minimum: 1
          default: 10
        is_active:
          type: boolean
          default: true
...
    ActionHistory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        device_type:
          type: string
          enum: [FEEDER, UV]
          example: "FEEDER"
        trigger_source:
          type: string
          enum: [SCHEDULE, MANUAL]
          example: "MANUAL"
        start_time:
          type: string
          format: date-time
          example: "2025-11-19T09:26:00Z"
        end_time:
          type: string
          format: date-time
          nullable: true
          example: "2025-11-19T09:26:05Z"
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, FAILED, STOPPED]
          example: "SUCCESS"
        value:
          type: integer
          description: |
            Nilai dalam satuan yang sesuai dengan device_type:
            - **FEEDER**: Jumlah pakan dalam gram (e.g., 10, 15, 20)
            - **UV**: Durasi dalam detik (e.g., 7200 untuk 2 jam, 28800 untuk 8 jam)

            Untuk UV schedule, value berisi remaining duration dari waktu schedule dimulai hingga selesai.
          example: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Stock:
      type: object
      properties:
        id:
          type: integer
          example: 1
        amount_gram:
          type: integer
          minimum: 0
          description: Jumlah stock dalam gram
          example: 1000
        updated_at:
          type: string
          format: date-time
    SensorLog:
      type: object
      properties:
        id:
          type: integer
          example: 1
        temperature:
          type: number
          format: float
          description: Temperature in Celsius
          example: 27.5
        humidity:
          type: number
          format: float
          description: Humidity percentage
          example: 65.0
        recorded_at:
          type: string
          format: date-time
          example: "2025-11-19T09:30:00Z"
...
    LastFeedInfo:
      type: object
      properties:
        day:
          type: string
          example: "Wednesday"
        time:
          type: string
          example: "08:00"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message here"
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20
        total:
          type: integer
          description: Total number of records
          example: 156
        total_pages:
          type: integer
          description: Total number of pages
          example: 8